requirements.txt
adal==1.2.7
antlr4-python3-runtime==4.9.3
applicationinsights==0.11.10
argcomplete==1.12.3
azure-appconfiguration==1.1.1
azure-applicationinsights==0.1.1
azure-batch==12.0.0
azure-cli==2.38.0
azure-cli-core==2.38.0
azure-cli-telemetry==1.0.6
azure-common==1.1.28
azure-core==1.24.1
azure-cosmos==3.2.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-data-tables==12.4.0
azure-datalake-store==0.0.52
azure-eventgrid==1.3.0
azure-graphrbac==0.60.0
azure-identity==1.10.0
azure-keyvault==1.1.0
azure-keyvault-administration==4.0.0b3
azure-keyvault-keys==4.5.1
azure-keyvault-secrets==4.4.0
azure-loganalytics==0.1.1
azure-mgmt==4.0.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==3.0.0
azure-mgmt-appconfiguration==2.1.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==2.0.0
azure-mgmt-batch==16.1.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0b3
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.2.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==27.1.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==9.1.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-containerservice==19.1.0
azure-mgmt-core==1.3.1
azure-mgmt-cosmosdb==7.0.0b6
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==9.0.0
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-hdinsight==9.0.0
azure-mgmt-imagebuilder==1.0.0
azure-mgmt-iotcentral==10.0.0b1
azure-mgmt-iothub==2.2.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==3.0.0
azure-mgmt-msi==6.0.1
azure-mgmt-netapp==8.0.0
azure-mgmt-network==20.0.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==1.1.0b2
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.0.0
azure-mgmt-recoveryservices==2.0.0
azure-mgmt-recoveryservicesbackup==5.0.0
azure-mgmt-redhatopenshift==1.1.0
azure-mgmt-redis==13.1.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==2.0.0
azure-mgmt-resource==21.1.0b1
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==8.0.0
azure-mgmt-security==2.0.0b1
azure-mgmt-servicebus==7.1.0
azure-mgmt-servicefabric==1.0.0
azure-mgmt-servicefabricmanagedclusters==1.0.0
azure-mgmt-servicelinker==1.0.0
azure-mgmt-signalr==1.0.0b2
azure-mgmt-sql==4.0.0b2
azure-mgmt-sqlvirtualmachine==1.0.0b3
azure-mgmt-storage==20.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-synapse==2.1.0b2
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==6.1.0
azure-multiapi-storage==0.9.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.13.0
azure-synapse-managedprivateendpoints==0.3.0
azure-synapse-spark==0.2.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.4
Deprecated==1.2.13
fabric==2.7.0
humanfriendly==10.0
idna==3.3
invoke==1.7.1
isodate==0.6.1
javaproperties==0.5.2
jmespath==1.0.1
jsondiff==2.0.0
knack==0.9.0
lxml==4.9.1
msal==1.18.0b1
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
multitasking==0.0.11
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
paramiko==2.11.0
pathlib2==2.3.7.post1
pkginfo==1.8.3
portalocker==1.7.1
psutil==5.9.1
pycparser==2.21
PyGithub==1.55
Pygments==2.12.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyodbc==4.0.32
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyreadline3==3.4.1
PySocks==1.7.1
python-certifi-win32==1.6
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
scp==0.13.6
semver==2.13.0
setuptools-scm==6.4.2
six==1.16.0
sshtunnel==0.1.5
tabulate==0.8.10
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
websocket-client==1.3.3
wrapt==1.14.1
xmltodict==0.13.0
